--- a/polylearn/cd_direct_fast.pyx
+++ b/polylearn/cd_direct_fast.pyx
@@ -10,8 +10,6 @@
 from libc.stdlib cimport malloc, free
 from libc.math cimport fabs
 
-cimport numpy as np
-
 from lightning.impl.dataset_fast cimport ColumnDataset
 
 from .loss_fast cimport LossFunction
@@ -19,7 +17,7 @@ from .cd_linear_fast cimport _cd_linear_epoch
 
 
 cdef void _precompute(ColumnDataset X,
-                      np.ndarray[double, ndim=3, mode='c'] P,
+                      double[:, :, ::1] P,
                       int order,
                       double* out,
                       int s,
@@ -45,18 +43,18 @@ cdef void _precompute(ColumnDataset X,
             out[i] += (data[ii] * P[order, s, j]) ** degree
 
 
-cdef double _update(int* indices,
-                    double* data,
-                    int n_nz,
-                    double p_js,
-                    np.ndarray[double, ndim=1] y,
-                    np.ndarray[double, ndim=1] y_pred,
-                    LossFunction loss,
-                    double* d1,
-                    double* d2,
-                    int degree,
-                    double lam,
-                    double beta):
+cdef inline double _update(int* indices,
+                           double* data,
+                           int n_nz,
+                           double p_js,
+                           double[:] y,
+                           double[:] y_pred,
+                           LossFunction loss,
+                           double* d1,
+                           double* d2,
+                           int degree,
+                           double lam,
+                           double beta):
 
     cdef double l1_reg = 2 * beta * fabs(lam)
     
@@ -90,54 +88,54 @@ cdef double _update(int* indices,
     return update
 
 
-cdef double _total_loss(np.ndarray[double, ndim=1] y,
-                        np.ndarray[double, ndim=1] y_pred,
-                        np.ndarray[double, ndim=3, mode='c'] P,
-                        np.ndarray[double, ndim=1] lams,
-                        np.ndarray[double, ndim=1] w,
-                        double alpha,
-                        double beta,
-                        LossFunction loss):
+# cdef double _total_loss(np.ndarray[double, ndim=1] y,
+#                         np.ndarray[double, ndim=1] y_pred,
+#                         np.ndarray[double, ndim=3, mode='c'] P,
+#                         np.ndarray[double, ndim=1] lams,
+#                         np.ndarray[double, ndim=1] w,
+#                         double alpha,
+#                         double beta,
+#                         LossFunction loss):
+#
+#     cdef double result = 0
+#     cdef double linear = 0
+#     cdef double row_norm
+#
+#     cdef int n_orders = P.shape[0]
+#     cdef int n_components = P.shape[1]
+#     cdef int n_features = P.shape[2]
+#     cdef int n_samples = y.shape[0]
+#
+#     cdef int i, s, j
+#
+#     # regularization
+#     for o in range(n_orders):
+#         for s in range(n_components):
+#             row_norm = 0
+#             for j in range(n_features):
+#                 row_norm += P[o, s, j] ** 2
+#             result += row_norm * lams[s]
+#     result *= beta
+#
+#     for j in range(n_features):
+#         linear += w[j] ** 2
+#
+#     linear *= alpha
+#
+#     result += linear
+#
+#     for i in range(n_samples):
+#         result += loss.loss(y_pred[i], y[i])
+#
+#     return result
 
-    cdef double result = 0
-    cdef double linear = 0
-    cdef double row_norm
 
-    cdef int n_orders = P.shape[0]
-    cdef int n_components = P.shape[1]
-    cdef int n_features = P.shape[2]
-    cdef int n_samples = y.shape[0]
-
-    cdef int i, s, j
-
-    # regularization
-    for o in range(n_orders):
-        for s in range(n_components):
-            row_norm = 0
-            for j in range(n_features):
-                row_norm += P[o, s, j] ** 2
-            result += row_norm * lams[s]
-    result *= beta
-
-    for j in range(n_features):
-        linear += w[j] ** 2
-
-    linear *= alpha
-
-    result += linear
-
-    for i in range(n_samples):
-        result += loss.loss(y_pred[i], y[i])
-
-    return result
-
-
-cdef double _cd_direct_epoch(np.ndarray[double, ndim=3, mode='c'] P,
+cdef inline double _cd_direct_epoch(double[:, :, ::1] P,
                              int order,
                              ColumnDataset X,
-                             np.ndarray[double, ndim=1] y,
-                             np.ndarray[double, ndim=1] y_pred,
-                             np.ndarray[double, ndim=1] lams,
+                             double[:] y,
+                             double[:] y_pred,
+                             double[:] lams,
                              double* d1,
                              double* d2,
                              int degree,
@@ -191,13 +189,13 @@ cdef double _cd_direct_epoch(np.ndarray[double, ndim=3, mode='c'] P,
     return sum_viol
 
 
-def _cd_direct_ho(np.ndarray[double, ndim=3, mode='c'] P,
-                  np.ndarray[double, ndim=1] w,
+def _cd_direct_ho(double[:, :, ::1] P,
+                  double[:] w,
                   ColumnDataset X,
-                  np.ndarray[double, ndim=1] col_norm_sq,
-                  np.ndarray[double, ndim=1] y,
-                  np.ndarray[double, ndim=1] y_pred,
-                  np.ndarray[double, ndim=1, mode='c'] lams,
+                  double[:] col_norm_sq,
+                  double[:] y,
+                  double[:] y_pred,
+                  double[:] lams,
                   int degree,
                   double alpha,
                   double beta,
@@ -236,9 +234,9 @@ def _cd_direct_ho(np.ndarray[double, ndim=3, mode='c'] P,
 
         if verbose:
             print("Iteration", it + 1, "violation sum", viol, end=" ")
-            if compute_loss:
-                print("loss", _total_loss(y_pred, y, P, lams, w, alpha, beta,
-                                          loss), end="")
+            # if compute_loss:
+            #     print("loss", _total_loss(y_pred, y, P, lams, w, alpha, beta,
+            #                               loss), end="")
             print()
 
         if viol < tol: