--- a/polylearn/cd_lifted_fast.pyx
+++ b/polylearn/cd_lifted_fast.pyx
@@ -10,8 +10,6 @@
 from libc.stdlib cimport malloc, free
 from libc.math cimport fabs
 
-cimport numpy as np
-
 from lightning.impl.dataset_fast cimport ColumnDataset
 
 from .loss_fast cimport LossFunction