--- a/polylearn/cd_linear_fast.pxd
+++ b/polylearn/cd_linear_fast.pxd
@@ -1,4 +1,3 @@
-cimport numpy as np
 from lightning.impl.dataset_fast cimport ColumnDataset
 from .loss_fast cimport LossFunction