… appveyor.yml
… benchmarks/bench_20newsgroups.py
… polylearn/cd_direct_fast.cpp
… polylearn/cd_lifted_fast.cpp
… polylearn/cd_linear_fast.cpp
… polylearn/factorization_machine.py
… polylearn/loss_fast.cpp
… polylearn/loss_fast.pyx
… polylearn/polynomial_network.py
… polylearn/tests/test_factorization_machine.py