Switch everything to typed memoryviews and inline direct solver update.

```
20 newsgroups
=============
X_train.shape = (11314, 130107)
X_train.format = csc
X_train.dtype = float64
X_train density = 0.001214353154362896
y_train (11314,)
X_test (7532, 130107)
X_test.format = csc
X_test.dtype = float64
y_test (7532,)

Classifier Training
===================
/home/vlad/code/polylearn/polylearn/factorization_machine.py:118: UserWarning: Objective did not converge. Increase max_iter.
  warnings.warn("Objective did not converge. Increase max_iter.")
Training fm-2 ... done
Training fm-3 ... done
/home/vlad/code/polylearn/polylearn/polynomial_network.py:99: UserWarning: Objective did not converge. Increase max_iter.
  warnings.warn("Objective did not converge. Increase max_iter.")
Training polynet-2 ... done
Training polynet-3 ... done
Classification performance:
===========================

Classifier            train       test         f1   accuracy
------------------------------------------------------------
fm-3               36.3608s    0.3642s     0.4009     0.9663
fm-2               31.7401s    0.1225s     0.6126     0.9740
polynet-3           6.6819s    0.0685s     0.6992     0.9796
polynet-2           5.9581s    0.0693s     0.7310     0.9807

>>> on branch speedup

Classifier            train       test         f1   accuracy
------------------------------------------------------------
fm-3               11.7805s    0.3608s     0.4009     0.9663
fm-2                7.1945s    0.1218s     0.6126     0.9740
polynet-3           6.5308s    0.0676s     0.6992     0.9796
polynet-2           5.9655s    0.0675s     0.7310     0.9807

```