@vene 2016-07-19
@fabianp could you give me a hand to figure out why I'm getting [these win32/py2 failures](https://ci.appveyor.com/project/vene/polylearn/build/1.0.29)? It looks like the `cd_direct_fast.pyx` solver has numerical issues in this configuration, but I can't pinpoint the cause.