@fabianp 2016-07-20
Hey. Went through the file cd_direct_fast.pyx but I don't know what can be causing such failures. Any chance you could make a minimal example for such failures?