@mblondel 2016-07-25
Since the win32 failures are independent of this PR, I think this PR should be merged, since it is a big improvement. I would also add a benchmark/ folder to help track the improvements over time.

Regarding the failures, could it be that the random kit behave differently on win32?