@fabianp 2016-07-25
If that is the case, adding l2 regularization and/or more iterations should
solve the problem

On Jul 25, 2016 2:49 PM, "Mathieu Blondel" <notifications@github.com> wrote:

> Since the win32 failures are independent of this PR, I think this PR
> should be merged, since it is a big improvement. I would also add a
> benchmark/ folder to help track the improvements over time.
>
> Regarding the failures, could it be that the random kit behave differently
> on win32?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/scikit-learn-contrib/polylearn/pull/1#issuecomment-234943596>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAQ8hwPuHbnZQYCovhlsgRQJNBkR_VrZks5qZLDDgaJpZM4JO8xX>
> .
>