@mblondel 2016-07-25
Remember that FMs and PNs are non-convex :)