@mblondel 2016-07-25
@ogrisel From past experience with scikit-learn, any idea what the win32 failures could come from?

By the way, the problem is only on Python 2.7. Python 3.5 is fine.