@fabianp 2016-07-25
D'oh!

On Jul 25, 2016 2:55 PM, "Mathieu Blondel" <notifications@github.com> wrote:

> Remember that FMs and PNs are non-convex :)
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/scikit-learn-contrib/polylearn/pull/1#issuecomment-234945037>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAQ8h5Z2gdf70Izpa9Lk0nEZLJJODZdHks5qZLJJgaJpZM4JO8xX>
> .
>