@ogrisel 2016-07-25
`Py_ssize_t` has a different meaning signed 32 bit int on a 32 bit platform and signed 64 bit int on a 64 bit platform. It should only be used to index datastructures (e.g. arrays or Python lists). I do not understand why the problem does not exist with Python 3 though.