@vene 2016-07-25
That makes sense. Thanks, @ogrisel! However the problem still existed when I was using int everywhere. I only replaced it with py_ssize_t in an attempt to fix it, and it seems I got a bit overzealous and replaced too many vars.

I'll change it to use py_ssize_t just where it should.

On July 25, 2016 10:23:48 AM EDT, Olivier Grisel <notifications@github.com> wrote:
>`Py_ssize_t` has a different meaning signed 32 bit int on a 32 bit
>platform and signed 64 bit int on a 64 bit platform. It should only be
>used to index datastructures (e.g. arrays or Python lists). I do not
>understand why the problem does not exist with Python 3 though.
>
>---
>You are receiving this because you authored the thread.
>Reply to this email directly or view it on GitHub:
>https://github.com/scikit-learn-contrib/polylearn/pull/1#issuecomment-234967790

-- 
Sent from my Android device with K-9 Mail. Please excuse my brevity.