@vene 2016-07-25
@mblondel I think it's not a rng issue. The `y` generated in `tests/test_factorization_machine.py:check_fit` for both `degree=2, 3` is the same in AppVeyor as on my 64bit ubuntu machine. I haven't directly checked the initial P_ but I doubt it would be different.

I also fixed the types as suggested by @ogrisel with no difference.

So it's probably something numerical.  It's strange that no errors occur with the lifted solver though.

Maybe I should try a 32bit py2 env in linux and see if I can reproduce it.