@mblondel 2016-07-26
Is there a way to reduce the verbosity of coveralls? I receive one email per commit, which makes it a bit difficult to follow real comments.