@vene 2016-08-01
I fixed the windows bug! It was my fault, of course. I forgot to initialize a `double` scalar to 0, and it turns out this doesn't behave well on MSVC (oddly, only for py2, not for py3: must be because of the different compiler version?!)

Appveyor is green now for both win32 and win64. (I canceled a duplicate build, triggered by the fact that this branch is on scikit-learn-contrib rather than on my fork, feel free to ignore that.)

I'll remove the ugly `_sq` thing, update the benchmarks, then ping again for a final review before merging.