@mblondel 2016-08-02
> I fixed the windows bug!

Hurray! Squash and merge?