@vene 2016-08-02
[Merged by rebase.](https://github.com/scikit-learn-contrib/polylearn/commit/195f4cd493564d18c7725acb40d9f341dda5bb0a) :beers: